backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
orjson==3.10.7
sqlalchemy==2.0.36
aiosqlite==0.20.0
PyJWT==2.9.0
python-multipart==0.0.12
Pillow==10.4.0
